+++ libspider/memory/dynamic-policies/FreeListAllocatorPolicy.h
@@ -71,7 +71,7 @@ public:
 
     void *allocate(size_t size) override;
 
-    size_t deallocate(void *ptr) override;
+    u64 deallocate(void *ptr) override;
 
     static size_t MIN_CHUNK_SIZE;
 private:
