+++ libspider/memory/dynamic-policies/GenericAllocatorPolicy.h
@@ -54,7 +54,7 @@ public:
 
     void *allocate(size_t size) override;
 
-    size_t deallocate(void *ptr) override;
+    u64 deallocate(void *ptr) override;
 };
 
 #endif //SPIDER2_GENERICALLOCATORPOLICY_H
